chore: update phpstan version and remove phpstan-baseline
This commit updates the version of phpstan specified in .phive/phars.xml file to 1.* and the
current installed version to 1.12.13.

Additionally, it removes all the previously ignored errors in phpstan-baseline.neon. The actions
were taken possibly due to fixes to these errors or related update in the codebase causing them to
be obsolete.

Please ensure related checks/errors do not surface again due to the removed baseline ignores.